added some items back
