+++ public/index.html
@@ -1,10 +1,10 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
-    <title>Redux Template</title>
-    <!-- <link href="/public/css/main.css" rel="stylesheet" /> -->
+    <title>dita</title>
+    <link href="/public/css/main.css" rel="stylesheet" />
     <link href="/vendor/bootstrap/dist/css/bootstrap.css" rel="stylesheet" />
-    <!-- <link href="/public/css/custom.css" rel="stylesheet" /> -->
+    <link href="/public/css/custom.css" rel="stylesheet" />
     <link rel="icon" href="/public/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
   </head>
